+++ libesp32/ESP32-to-ESP8266-compat/src/c_types.h
@@ -1,0 +1,6 @@
+#pragma once
+/**/
+#include <stdint.h>
+#ifndef ICACHE_FLASH_ATTR
+#define ICACHE_FLASH_ATTR
+#endif