+++ tasmota/tasmota_template.h
@@ -226,6 +226,7 @@ enum UserSelectablePins {
   GPIO_CC1101_GDO2,    // CC1101 pin for RX
   GPIO_HRXL_RX,       // Data from MaxBotix HRXL sonar range sensor
   GPIO_ELECTRIQ_MOODL_TX, // ElectriQ iQ-wifiMOODL Serial TX
+  GPIO_AS3935,
   GPIO_SENSOR_END };
 
 // Programmer selectable GPIO functionality
@@ -312,7 +313,8 @@ const char kSensorNames[] PROGMEM =
   D_SENSOR_LE01MR_RX "|" D_SENSOR_LE01MR_TX "|"
   D_SENSOR_CC1101_GDO0 "|" D_SENSOR_CC1101_GDO2 "|"
   D_SENSOR_HRXL_RX "|"
-  D_SENSOR_ELECTRIQ_MOODL
+  D_SENSOR_ELECTRIQ_MOODL "|"
+  D_SENSOR_AS3935 
   ;
 
 const char kSensorNamesFixed[] PROGMEM =
@@ -656,6 +658,9 @@ const uint8_t kGpioNiceList[] PROGMEM = {
 #ifdef USE_HRXL
   GPIO_HRXL_RX,
 #endif
+#ifdef USE_AS3935
+  GPIO_AS3935,
+#endif
 };
 
 /********************************************************************************************/
@@ -692,10 +697,25 @@ const char kAdc0Names[] PROGMEM =
 
 /********************************************************************************************/
 
+#ifdef ESP8266
+
 #define MAX_GPIO_PIN       17   // Number of supported GPIO
 #define MIN_FLASH_PINS     4    // Number of flash chip pins unusable for configuration (GPIO6, 7, 8 and 11)
 
 const char PINS_WEMOS[] PROGMEM = "D3TXD4RXD2D1flashcFLFLolD6D7D5D8D0A0";
+
+#else  // ESP32
+
+// esp32 has more pins
+#define USER_MODULE        255
+#define MAX_GPIO_PIN       44   // Number of supported GPIO
+#define MIN_FLASH_PINS     4    // Number of flash chip pins unusable for configuration (GPIO6, 7, 8 and 11)
+
+const char PINS_WEMOS[] PROGMEM = "00010203040506070809101112131415161718192021222324252627282930313233343536373839";
+
+#endif  // ESP8266
+
+/********************************************************************************************/
 
 typedef struct MYIO {
   uint8_t      io[MAX_GPIO_PIN];
@@ -729,6 +749,8 @@ typedef struct MYTMPLT {
 } mytmplt;
 
 /********************************************************************************************/
+
+#ifdef ESP8266
 
 // Supported hardware modules
 enum SupportedModules {
@@ -2220,5 +2242,11 @@ const mytmplt kModules[MAXMODULE] PROGMEM = {
     0, 0, 0, 0
   }
 };
+
+#endif  // ESP8266
+
+#ifdef ESP32
+#include "tasmota_template_ESP32.h"
+#endif  // ESP32
 
 #endif  // _TASMOTA_TEMPLATE_H_